refactor(architecture): move files under project
