+++ src/core/project.rs
@@ -7,6 +7,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 pub mod component;
+pub mod files;
 mod readiness;
 mod status;
 
@@ -16,6 +17,7 @@ pub use component::{
     resolve_project_component, resolve_project_components,
     set_component_attachments,
 };
+pub use files::{FileEntry, GrepMatch, LineChange};
 pub use readiness::calculate_deploy_readiness;
 pub use status::{collect_status, ProjectComponentStatus, ProjectStatusSnapshot};
